# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  dtos.AccountBalanceDTO:
    properties:
      balance:
        type: number
      currency_code:
        type: string
      currency_symbol:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      type:
        type: string
    type: object
...
  dtos.AccountSummary:
    properties:
      balance:
        description: Can be string or decimal.Decimal depending on model version
      currency:
        $ref: '#/definitions/dtos.CurrencySummary'
      id:
        type: integer
      name:
        type: string
      type:
        type: string
    type: object
  dtos.BalanceVerificationResponse:
    properties:
      difference:
        description: Should be 0 for balanced transactions
        type: number
      is_balanced:
        type: boolean
      message:
        type: string
      total_credits:
        type: number
      total_debits:
        type: number
      transaction_id:
        type: integer
    type: object
...
  dtos.CurrencySummary:
    properties:
      code:
        type: string
      id:
        type: integer
      symbol:
        type: string
    type: object
  dtos.DashboardResponse:
    properties:
      account_balances:
        description: Account Breakdown
        items:
          $ref: '#/definitions/dtos.AccountBalanceDTO'
        type: array
      as_of:
        description: Metadata
        type: string
      average_monthly_expenses:
        description: Last 3 months average
        type: number
      base_currency:
        description: 'User''s base currency (default: USD)'
        type: string
      liquid_assets:
        description: BANK + CASH only
        type: number
      monthly_expenses:
        description: Total EXPENSE transactions this month
        type: number
      monthly_income:
        description: Monthly Stats (current month)
        type: number
      monthly_net_cash_flow:
        description: Income - Expenses
        type: number
      net_worth:
        description: Assets - Liabilities
        type: number
      runway:
        description: Runway Calculation (Critical feature from DOC.md)
        type: number
      runway_days:
        description: Runway in days
        type: integer
      total_assets:
        description: Financial Overview
        type: number
      total_liabilities:
        description: Sum of all CREDIT_CARD accounts
        type: number
    type: object
  dtos.ErrorResponse:
    properties:
      details:
        example: additional error details
        type: string
      error:
        example: error message
        type: string
    type: object
  dtos.JournalEntryListResponse:
    properties:
      entries:
        items:
          $ref: '#/definitions/dtos.JournalEntryResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  dtos.JournalEntryResponse:
    properties:
      account:
        allOf:
        - $ref: '#/definitions/dtos.AccountSummary'
        description: Relationships
      account_id:
        type: integer
      amount:
        type: number
      created_at:
        type: string
      debit_or_credit:
        description: '"DEBIT" or "CREDIT"'
        type: string
      description:
        type: string
      entry_date:
        type: string
      id:
        type: integer
      transaction_id:
        type: integer
      user_id:
        type: integer
    type: object
...
  dtos.MonthlyStats:
    properties:
      expenses:
        type: number
      income:
        type: number
      month:
        description: 1-12
        type: integer
      net_cash_flow:
        description: Income - Expenses
        type: number
      transaction_count:
        type: integer
      year:
        type: integer
    type: object
...
  dtos.RunwayCalculation:
    properties:
      available_funds:
        description: LiquidAssets - ShortTermLiabilities
        type: number
      average_monthly_expenses:
        description: Last 3 months
        type: number
      bank_accounts:
        description: Breakdown by account type
        items:
          $ref: '#/definitions/dtos.AccountBalanceDTO'
        type: array
      base_currency:
        type: string
      calculation_date:
        type: string
      cash_accounts:
        items:
          $ref: '#/definitions/dtos.AccountBalanceDTO'
        type: array
      credit_card_accounts:
        items:
          $ref: '#/definitions/dtos.AccountBalanceDTO'
        type: array
      liquid_assets:
        description: BANK + CASH
        type: number
      message:
        type: string
      runway_days:
        type: integer
      runway_months:
        description: AvailableFunds / AvgExpenses
        type: number
      short_term_liabilities:
        description: CREDIT_CARD balances
        type: number
      status:
        description: Warning levels
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@arabella.app
    name: Arabella Support
  description: API REST para el sistema de gestión financiera personal Arabella. Implementa
    contabilidad de doble partida, cálculo de Runway y soporte multi-moneda.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Arabella Financial OS API
  version: 1.0.0
paths:
...
  /dashboard:
    get:
      description: 'Obtiene el dashboard completo del usuario autenticado: patrimonio
        neto, activos, pasivos, flujo de caja mensual, y el cálculo de Runway (cuántos
        meses puede sobrevivir sin ingresos basado en el promedio de gastos de los
        últimos 3 meses)'
      produces:
      - application/json
      responses:
        "200":
          description: Dashboard financiero completo
          schema:
            properties:
              data:
                $ref: '#/definitions/dtos.DashboardResponse'
            type: object
        "401":
          description: No autenticado
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Resumen financiero del usuario
      tags:
      - Dashboard
  /dashboard/monthly-stats:
    get:
      description: Obtiene el resumen de ingresos, gastos y flujo neto de caja para
        un mes y año específicos del usuario autenticado
      parameters:
      - description: 'Mes (1-12, default: mes actual)'
        in: query
        name: month
        type: integer
      - description: 'Año (ej: 2026, default: año actual)'
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Estadísticas del mes solicitado
          schema:
            properties:
              data:
                $ref: '#/definitions/dtos.MonthlyStats'
            type: object
        "401":
          description: No autenticado
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Estadísticas mensuales de ingresos y gastos
      tags:
      - Dashboard
  /dashboard/runway:
    get:
      description: 'Calcula cuántos meses y días puede el usuario sostenerse financieramente
        sin nuevos ingresos. Fórmula: (Activos líquidos - Pasivos a corto plazo) /
        Promedio de gastos mensuales (últimos 3 meses). Estados: HEALTHY (≥6 meses),
        WARNING (3-6 meses), CRITICAL (<3 meses)'
      produces:
      - application/json
      responses:
        "200":
          description: Cálculo de Runway detallado con desglose por tipo de cuenta
          schema:
            properties:
              data:
                $ref: '#/definitions/dtos.RunwayCalculation'
            type: object
        "401":
          description: No autenticado
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cálculo detallado de Runway
      tags:
      - Dashboard
...
  /journal-entries:
    get:
      description: Obtiene una lista paginada de asientos del diario contable del
        usuario autenticado. Los asientos son generados automáticamente por el Motor
        Contable al crear o revertir transacciones; nunca se crean ni modifican manualmente.
        Se pueden filtrar por transacción, cuenta o tipo (DEBIT/CREDIT)
      parameters:
      - description: Filtrar por ID de transacción
        in: query
        name: transaction_id
        type: integer
      - description: Filtrar por ID de cuenta
        in: query
        name: account_id
        type: integer
      - description: Filtrar por tipo de asiento (DEBIT, CREDIT)
        in: query
        name: debit_or_credit
        type: string
      - description: 'Número de página (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Elementos por página (default: 50, máx: 100)'
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lista paginada de asientos contables
          schema:
            $ref: '#/definitions/dtos.JournalEntryListResponse'
        "401":
          description: No autenticado
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar asientos contables (pista de auditoría)
      tags:
      - Journal Entries
  /journal-entries/transaction/{id}:
    get:
      description: Obtiene todos los asientos de débito y crédito generados para una
        transacción específica. Cada transacción produce exactamente 2 asientos (uno
        DEBIT y uno CREDIT) que deben sumar igual, garantizando el equilibrio de la
        doble partida
      parameters:
      - description: ID de la transacción
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Asientos de la transacción
          schema:
            properties:
              count:
                type: integer
              data:
                items:
                  $ref: '#/definitions/dtos.JournalEntryResponse'
                type: array
            type: object
        "400":
          description: ID inválido
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "401":
          description: No autenticado
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Asientos contables de una transacción
      tags:
      - Journal Entries
  /journal-entries/verify/{id}:
    get:
      description: 'Verifica que la suma de todos los asientos DEBIT sea igual a la
        suma de los asientos CREDIT para una transacción dada. Esto garantiza la regla
        fundamental de la contabilidad de doble partida: ∑ Débitos = ∑ Créditos'
      parameters:
      - description: ID de la transacción a verificar
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Resultado de la verificación con totales de débito y crédito
          schema:
            properties:
              data:
                $ref: '#/definitions/dtos.BalanceVerificationResponse'
            type: object
        "400":
          description: ID inválido
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "401":
          description: No autenticado
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Error al ejecutar la verificación
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Verificar equilibrio contable de una transacción
      tags:
      - Journal Entries
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Escribe "Bearer" seguido de un espacio y tu JWT de acceso. Ejemplo:
      "Bearer eyJhbGci..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"